--- server/src/services/handleRecipeForm.js
@@ -1,11 +1,0 @@
-import { Recipe } from '../models/index.js'
-
-const handleRecipeForm = async (req) => {
-  const { name, image, url, source } = req.body
-  const ingredients = req.body.ingredients.map(({ name }) => name).join('\n')
-  const instructions = req.body.instructions.map(({ name }) => name).join('\n')
-  const newRecipe = await Recipe.query().insertAndFetch({ name, ingredients, instructions, image, url, source })
-  await newRecipe.$relatedQuery('users').relate( req.user.id )
-}
-
-export default handleRecipeForm
