+++ server/src/services/pythonScraper.js
